Almost done. Started work on websockets.

+++ Backend/controllers/room.controller.ts
@@ -1,9 +1,17 @@
 import Room from '../models/room.model.ts';
 import Booking from '../models/booking.model.ts';
 import User from '../models/user.model.ts';
+import { logger } from '../utils/logger.utils.ts';
 import { Request, Response } from 'express';
 
 export const createRoom = async(req: Request, res: Response) : Promise<void> => {
+    
+    if (!req.user) {
+        logger.error(`Something went wrong when finding the user.`)
+        res.status(401).json({ message: "Unauthorized" });
+        return;
+    }
+
     try{
         const {
             address, 
@@ -16,8 +24,9 @@ export const createRoom = async(req: Request, res: Response) : Promise<void> => 
         } = req.body;
         
         const userObject = await User.findOne({username: name})
+
         if (userObject?.roleRaise === false) {
-            console.log("The user has not been confirmed")
+            logger.error(`The users application has not yet been accepted.`)
             res.status(403).send()
             return
         }
@@ -34,41 +43,66 @@ export const createRoom = async(req: Request, res: Response) : Promise<void> => 
 
         await newRoom.save()
 
+        logger.error(`The room has been created and added to the database.`)
         res.status(201).json(newRoom)
     } catch(error) {
-        console.error('Error creating room:', error);
+        logger.error(`There was an error on the server side.`)
         res.status(500).json({ message: 'Server error' })
     }
 }
 
 export const getRooms = async(req: Request, res: Response) :  Promise<void> => {
+    
+    if (!req.user) {
+        logger.error(`There seems to be something wrong with the `)
+        res.status(401).json({ message: "Unauthorized" });
+        return;
+    }
+
     try {
         const rooms = await Room.find({}).lean();
-        console.log(`Getting the rooms: \n${rooms}`)
+        logger.info(`The rooms are being sent back to the frontend.`)
         res.status(200).json(rooms);
     } catch(error) {
-        console.error('Error fetching rooms:', error);
+        logger.error(`Something went wrong when trying to fetch the rooms.`)
         res.status(500).json({ error: 'An error occurred while fetching rooms.' });
     }
 }
 
 export const removeRoom = async(req: Request, res: Response) : Promise<void> => {
+    
+    if (!req.user) {
+        logger.error(`The user could not be found.`)
+        res.status(401).json({ message: "Unauthorized" });
+        return;
+    }
+
     try {
         const { id } = req.params;
-        const deletedBookings = await Booking.deleteMany({roomId: id});
-        console.log(`Succesfully deleted the following bookings: \n${deletedBookings}`)
+        await Booking.deleteMany({roomId: id});
         const room = await Room.findByIdAndDelete(id);
         if (!room) {
-          res.status(404).json({ message: 'Room not found' });
+            logger.error(`Could not find the room.`)
+            res.status(404).json({ message: 'Room not found' });
+            return
         }
+        logger.info(`The room was succesfully deleted.`)
         res.status(200).json({ message: 'Room deleted successfully' });
     } catch(error) {
-        console.error(error);
+        logger.error(`There was an error during the deletion function.`)
         res.status(500).json({ message: 'Deletion Error'})
     }
 }
 
 export const updateRoom = async(req: Request, res: Response) : Promise<void> => {
+
+    if (!req.user) {
+        logger.error(`Could not find the user.`)
+        res.status(401).json({ message: "Unauthorized" });
+        return;
+    }
+    
+
     try {
         const {id} = req.params;
         const {
@@ -83,13 +117,10 @@ export const updateRoom = async(req: Request, res: Response) : Promise<void> => 
 
         const userObject = await User.findOne({username: name})
         if (userObject?.roleRaise === false) {
-            console.log("The user has not been confirmed")
+            logger.error(`The application has not yet been accepted.`)
             res.status(403).send()
             return
         }
-        
-        console.log(`This was the ID: ${id} \n\n This was the Payload: \n ${address}\n${name}\n${roomNumber}\n${roomOpens}\n${roomCloses}\n${capacity}\n${type}\n`);
-        
         const updatedRoom = await Room.findByIdAndUpdate(
             id,
             { address, name, roomNumber, roomOpens, roomCloses, capacity, type },
@@ -97,13 +128,15 @@ export const updateRoom = async(req: Request, res: Response) : Promise<void> => 
         );
       
         if (!updatedRoom) {
+            logger.error(`Could not find the room.`)
             res.status(404).json({ message: 'Room not found' });
             return;
         }
 
+        logger.info(`The room was succesfully updated.`)
         res.status(204).send()
     } catch(error) {
-        console.error(error);
+        logger.error(`There was an updated error on the server side.`)
         res.status(500).json({ message: 'Update Error'})
     }
 }
